- Added support for implicit block calls.
Example:
	f(a,b) {...};   # means: f(a, b, {...});
	f {...};        # means: f({...});

This new syntax allows the user to extend the language freely, creating new constructs using blocks, classes, methods and functions.
See: `scripts/block_implicit_call.sf` for examples.

However, this new feature comes at a cost, which is that it will always require parentheses around if/while/for conditions.

Example:
	if true   { ... };    # this used to work
	if (true) { ... };    # this works

Anyway, the parentheses make the code much easier to read and to reason about, so I think it's a good change of the syntax. It, also, makes the language more consistent.