- Added optional type-checking for function and method parameters.
Example:
	func add(Number a, Number b) {   # requires two numbers
	    a + b
	}
	say add(1, 2);		# prints: 3
	say add("a", "b");	# run-time error: invalid type for parameters

The old syntax (which was "a is Number, b is Number") has been removed and will fail to compile.

The new syntax also allows default values for typed paramters:

	func add(Number a=2, Number b=3) {
	    a + b
	}
	say add();		# prints: 5
	say add(b:8);		# pritns: 10

- Improved the type-checking of return types. Inheritance is now checked correctly.

Example:
	func add(a, b) -> Num {
	   a + b
	}
	say add(1, 2);		# prints: 3
	say add("a", "b");      # run-time error: invalid return type

In addition, by specifying "Object" instead of "Num", it will work with any object.
(this is true for function paramaters as well)