- Special methods `==`, `<=>` and `~~` are now overloadable.
This means that we are now allowed to redefine them in any user-defined class and the things will just work as expected.

For an example, please see: scripts/Applications/Pure OO/numbers.sf

As a side-effect, `~~` is now slightly faster, `==` is slightly slower, while `<=>` is about the same (or maybe a little bit slower). However, the overall trade-off is way too good and this feature also makes the language more consistent and dynamic.