- Added precompilation support for Sidef scripts (with option -s).
This technique is meant to reduces the boot-time, by saving each compiled code in a DBM, which is updated automatically and sanitized periodically.
Code stored inside the database is compressed with DEFLATE, which reduces its size significantly, with also no-cost at decompression.

This also led to a more user-friendly interface to the Parser and to the backend deparsers, by creating a Sidef object as:

	my $obj = Sidef->new(...);

Compiling a Sidef script is as simple as:

	$obj->compile_code(\'say "hi"', 'Perl');