- Better implicit numeric conversions.
The new method allows any object to be converted to a number, if it defines the `to_n` method.

Example:
	42 + obj		# means: `42 + obj.to_n`

When the method `to_n` is not defined, it will fail with an appropriate error message.

The same change is now also available for Complex objects, by calling the `to_c` method (if defined), falling back to method `to_n`.

Example:
	42 + obj		# means: `42 + obj.to_c`

Adding a number with an array is now also possible:

	42 + ['a', 'b', 'c']	# means: 42 + ['a','b','c'].len

- Added the Hash.pair(key) method, which returns a Pair object.

Example:
	var h = Hash(a => 1, b => 2)
	say h.pair(:a)			# prints: Pair('a', 1)

- Added the Hash.pairs(keys...) method, to return an array of Pair objects.

Example:
	var h = Hash(a => 1, b => 2)
	say h.pair(:b, :a)		# prints: [Pair('b',2),Pair('a',1)]