- Values returned by a Perl library are no longer converted to Sidef when the result is not needed (that is when the method is invoked in void-context).
In addition, both module-call managers, `Module::Func` and `Module::OO`, are now aware of the context where the result is needed. For example, if a method returns a list of values, but the method is called in scalar context, the list is now automatically converted into an object Array. This makes working with Perl modules a little bit more transparent.