- Added the `Number.cfrac()` method, which returns an Array that contains the continued-fraction representation of a given number.
Example:

	say as_cfrac(365.2422)	#=> [365, 4, 7, 1, 3, 4, 1, 1, 1, 2]

A continued-fraction returned by `Number.as_cfrac()` can be converted back to a number using the `Array.reduce{}` method:

	var cf = as_cfrac(365.2422)
	say cf.reverse.reduce { |a,b| b + 1/a }	  #=> 365.2422

This method is also defined to floating-point and complex numbers, based on the current numerical precision. Additionally, the method accepts an optional argument that specified the number of terms to be returned.

Example:

	say as_cfrac(Num.pi, 5)	  #=> [3, 7, 15, 1, 292]