- Added the Number `faulhaber_sum(n, p)` method. (also aliases as Number `faulhaber(n, p)`)
Example:

	say (faulhaber(10000, 20) == sum(1..10000, {_**20}))		#=> true

This function uses Faulhaber's formula to the sum of the first `n` natural numbers, each raised to the power `p`.

Example:

	faulhaber(5, 2) 	# means: 1**2 + 2**2 + 3**2 + 4**2 + 5**2

For large values of `n`, this functions is much more efficient that simply adding all the numbers together. More precisely, the function is liniar in `p` (i.e. it has a time complexity of O(p)).