- Reimplemented the `Number.digits()` method to support arbitrary large bases.
Example:

	say digits(1e9, 127)		#=> [95, 15, 24, 107, 3]

This matches the `.digits()` method from Ruby.

The previous implementation, when a base was specified, was equivalent with `n.base(b).chars`, which returned an array of strings, and `b` was limited to [2, 36]. In the current implementation, `n` and `b` can be arbitrary large and the method will always return an array of numbers, given that `n` and `b` are valid. When `abs(b)` is 1 or 0, it returns `nil`.