Reimplemented the `Number.digit()` method for better performance and to support arbitrary large bases.
Example:

	n.digit(i, b)	== n.digits(b)[i]

Where `n` is an integer, `b` is the base and `i` is the position of the digit (zero based and counting from the end, such as `i=0` means the last digit of `n`).

Exmaples:

	say 1119.digit(0)	  #=> 9
	say 1181.digit(1)	  #=> 8
	say 1711.digit(2)	  #=> 7
	say 6111.digit(3)	  #=> 6
	say 1234.digit(4)	  #=> 0

It also supports negative indices:

	say 9111.digit(-1)	  #=> 9
	say 1811.digit(-2)	  #=> 8
	say 1234.digit(-42)	  #=> nil

The new implementation is much faster when only one particular digit of a number is needed, specially when the number is very large.