- Added the Number `numify` method, which returns a raw native number representation for the self-number.
Can be used for assigning values to `Num!PREC` variable.

Example:

	local Num!PREC = 42.numify	# set precision to 42 bits
	say sqrt(2)			# 1.414213562

Setting precision to `42` directly (where `42` is a Number object), it's possible, but may be a little bit innefficient because the value has to be converted internally to a native number for each floating-point computation.

Native numbers can also be used when indexing an array:

	var arr = [42, 43, 44]
	var idx = 2.numify
	say arr[idx]			#=> 44

Although this may or may not be actually faster.