- Added the Number `primality_pretest(n)` method.
The method returns true when `n` passes the primality pretest (true if n is large enough (n > 2^15000) and it has no small factors less than 10^6).

If `n` is larger than 2^20000, the method does trial division up to 10^7.
If `n` is larger than 2^30000, the method does trial division up to 10^8.

This method can be used when multiple values have to be prime simultaneously. If one of the terms has a small factor, we don't need to run a primality test on any term (which is quite expensive when n is large enough).

Example:

	var p = 1e10000.irand
	var q = 1e10000.irand

	if (p.primality_pretest && q.primality_pretest && p.is_prime && q.is_prime) {
		say "Both p and q are primes"
	}
	else {
		say "Either p or q is composite (or both)"
	}