- Added the Array `.next_permutation` method.
It modifies the self array in place to contain the next unique permutation and returns true if there are more permutations available, or false of the current permutation is the last one.

Example:

	var array = [1,1,2]

	do {
		say array
	} while (array.next_permutation)

Output:

	[1, 1, 2]
	[1, 2, 1]
	[2, 1, 1]

- Also added the Array `.unique_permutations` method.

It uses the `next_permutation` method to create all the unique permutations.

Example:

	say [1,1,2].unique_permutations		#=> [[1, 1, 2], [1, 2, 1], [2, 1, 1]]

Equivalent with `array.permutations.uniq`, but more efficient, as it creates the permutations without duplicates.

The method also accepts a callback block as an optional argument:

	[1,1,2].unique_permutations {|*perm|
		say perm
	}

Output:

	[1, 1, 2]
	[1, 2, 1]
	[2, 1, 1]

See also:
	https://rosettacode.org/wiki/Permutations_with_some_identical_elements