- Added the Number `all_prime(...)` and `all_composite(...)` methods. - Also added the Array `.all_prime` and `.all_composite` methods, which call the corresponding Number methods.
The method `all_prime(a,b,c,...)` returns true if all the given values are prime numbers. This is done by first running a primality pre-test on all values and returning early if one of the values is composite with a small factor. Otherwise, we run a BPSW primality test on each value.
Similarly, the method `all_composite(a,b,c,...)` returns true if all the given values are composite positive integers, by checking first for small factors, then running a BPSW primality test.