- Minor simplifications inside the Optimizer.
No longer make DataTypes childs of the actual Types. Instead call the Types methods directly as a class-method.

This allows Num.pi to be called as Sidef::Types::Number::Number->pi, instead of bless({}, 'Sidef::DataTypes::Number::Number')->pi, allowing to check if the method is called as a class-method (i.e.: ref($_[0]) eq '') or as an object-method (i.e.: ref($_[0]) ne ''), and returing different results based on this information.

For example:

	say pi(100)		#=> 25 (number of primes <= 100)
	say Num.pi		#=> 3.14159... (the PI constant)