- Performance improvements in Number `is_lucas_carmichael(n)`.
If `lucas_factor()` finds some prime factors and some composites factors, we now check the prime factors for p+1 | n+1, so we can return early if the condition is not met, without factoring the composite factors. This is much faster for large `n` that are not Lucas-Carmichael.

The trial division is also improved, checking for small factors first and stopping early if `n` has no small factors (<= 5*10^4).