- Added the Polynomial `divmod(x,y)` method.
Returns the quotient and the remainder when Polynomial `x` is divided by Polynomial `y`.

- Added the Polynomial `self.coeff(e)` method.

Returns the coefficient for power `x^e`.

- Added the Polynomial `self.coeffs` method.

Returns a 2D Array with the exponents and the coefficients as [[e_1, c_1], [e_2, c_2], ..., [e_k, c_k]]

- Added the Polynomial `self.keys` method.

Returns an Array with the exponents `e`. To get the coefficient for each power, use `self.coeff(e)`.