- Added the Array `random_permutation` method.
Returns a random permutation of the self array.

This is equivalent with the `shuffle` method. It differs just in the implementation details: `random_permutation` uses the Number `irand` method for computing a random integer in the range `1..factorial(arr.len)`, then calls Array `nth_permutation` with this random value, while Array `shuffle` uses `List::Util::shuffle()` to shuffle the array, which uses perl's builtin `rand()` function.

The randomness quality of `random_permutation` is expected to be better than that of `shuffle`, although `random_permutation` is slighly slower.

Additionally, by setting a seed value for `irand` with Number `iseed()`, the results returned by `random_permutation` can be reproduced.

Example:

	var arr = %w(a b c d e f g)
	say arr.random_permutation

Example with iseed():

	iseed(42)
	var arr = %w(a b c d e f g)
        say arr.random_permutation	#=> ["d", "f", "g", "b", "c", "e", "a"]