- Added the `Math.product_tree(list...)` method.
Return the product-tree of a list of integers.

Example:

	Math.product_tree(10,20,30,40,50,60)

Output:

	[[10, 20, 30, 40, 50, 60], [200, 1200, 3000], [240000, 3000], [720000000]]

- Added the `Math.remainders(n, array)` method.

Efficiently returns the remainders of n when divided by each integer from the given array.

Example:

	say Math.remainders(8675309, [11,13,17,19,23])

Output:

	[5, 6, 5, 4, 8]

- Added the `Math.batch_gcd(list...)` method.

Efficiently compute the batch-gcd of a list of integers.

Example:

	Math.batch_gcd(1909,2923,291,205,989,62,451,1943,1079,2419)

Output:

	[1909, 1, 1, 41, 23, 1, 41, 1, 83, 41]

- Added the `Math.batch_invmod(array, n)` method.

Efficiently computes the modular multiplicative inverses of a list of integers, modulo n.

Example:

	say Math.batch_invmod([33, 42, 99, 103], 2017)

Output:

	[489, 1969, 163, 235]