- Improved the Optimizer to do constant folding on unary operators.
Additionally, it supports the creation of several more objects, such as Gauss, Quadratic, etc, when the arguments are literal values.

- Sidef parser: no longer parse `-` as part of a number.

Fixes the following issue:

	say -42**2	#=> -1764 (previously it printed "1764", because "-42" was parsed as a number)

Now `-` is parsed as a prefix unary operator, even when it is followed by a number.

This is now consistent with:

	var x = 42
	say -x**2	#=> -1764 (same output as before)

- Added the Range `neg` method.

It returns a new range with the starting point negated.

Example:

	say (-(3..4))	#=> RangeNum(-3, 4, 1)