- Optimized Number `sum(...)` for native integers.
Tries to sum native integers as a native sum, and falls-back to mpz when the size of the sum exceeds the size of a native integer.

However, the mpz object is now created on-demand (when used). Therefore, the method has now less overhead than before when called with native integers and the sum fits inside a native integer.

- Two simplifications in Number `__mod__(n,m)` when `n` is a Math::GMPz object, and `m` is a native positive integer.

- Make `RAT % INT` to be consistent with `INT % INT` when the modulo is negative.