- Added the Array `arr.ordered_partitions(k, { ... })` method.
Iterates over the ordered partitions `P` of `arr`, such that `P.flat == arr`.

Example:

	[1,2,3,4,5].ordered_partitions(2, {|*a| say a })

Outputs:
	[[1], [2, 3, 4, 5]]
	[[1, 2], [3, 4, 5]]
	[[1, 2, 3], [4, 5]]
	[[1, 2, 3, 4], [5]]

When no block is given, it returns an array with the ordered partitions (there are `binomial(n-1, k-1)` ordered partitions, where `n` is the size of the array):

	say [1,2,3,4].ordered_partitions(2) 	#=> [[[1], [2, 3, 4]], [[1, 2], [3, 4]], [[1, 2, 3], [4]]]

Additionally, when the value for `k` is ommited, it generates all ordered partitions (there are `2^(n-1)` ordered partitions, where `n` is the size of the array):

	[1,2,3,4].ordered_partitions{|*a| say a }

Outputs:
	[[1], [2], [3], [4]]
	[[1, 2], [3], [4]]
	[[1], [2, 3], [4]]
	[[1, 2, 3], [4]]
	[[1], [2], [3, 4]]
	[[1, 2], [3, 4]]
	[[1], [2, 3, 4]]
	[[1, 2, 3, 4]]

Up to ordering, the method is equivalent with (but more efficient):

	arr.partitions(k).grep { .flat == arr }

Update.