- Fixed a rare issue in Number `is_almost_prime(n,k)` and `is_omega_prime(n,k)`.
It occured when `special_factor` found one or more prime factors, but the composite part was still too large to be factorized (> 2^150), which resulted in a another trial-division step with a larger bound on the remainder based on the factors from `special_factor`. When the trial-division failed to find the factor found by `special_factor`, it overwrote the number of prime factors found so far, making the functions to return false-negatives for certain inputs.

And another rare issue when `special_factor` finds a prime factor, but the composite part is still too large to factor, then the `special_factor` is run again on the remainder, but lost the factor found earlier, which could also made the functions to return false-negative for certain inputs.

Both issues are now fixed.