- Added the Number `totient_range(a,b)` method.
Returns an array with the totient values for the range `a..b`.

- Added the Number `proper_sigma0(n)` method.

Returns the number of proper divisors of `n`.

- Added the Number `proper_divisors(n)` method.

Return an array with the divisors of `n` less than `n`.

- Added the Number `mobius_range(n)` method.

Returns an array with the Möbius values for the range `a..b`.

- Added the Number `is_deficient(n)` method.

Returns true if `n` is a deficient number, else false.

- Added the Number `is_amicable(n,m)` method.

Returns true if the numbers `n` and `m` are "amicable", else false.

- Extended the Number `aliquot(n)` method to accept an extra argument.

Example:

	say aliquot(n, 2)	#=> sum of proper divisors of n, each divisor squared
	say aliquot(n, -2)	#=> sum of proper divisors of n, each divisor raised to power -2