- Added the Number `znlog(a, g, m)` method.
Returns the integer `k` that solves the congruence: `a = g^k (mod m)`.

Example:

	say znlog(5675,5,10000019)	#=> 2003974 (where 5675 == 5^2003974 mod 10000019)

Returns `NaN` if a solution is not found.

The baby-step/giant-step algorithm is used, which uses O(sqrt(m)) space.

If Math::Prime::Util is installed, we use that instead for native integers, as it has a much more efficient algorithm for computing znlog().